Added RI reporting for TM3 and TM4 (no PMI)

+++ lib/include/srslte/phy/ue/ue_dl.h
@@ -190,7 +190,7 @@ SRSLTE_API int srslte_ue_dl_ri_pmi_select(srslte_ue_dl_t *q,
                                           float *current_sinr);
 
 SRSLTE_API int srslte_ue_dl_ri_select(srslte_ue_dl_t *q,
-                                      uint32_t *ri,
+                                      uint8_t *ri,
                                       float *cn);
 
 SRSLTE_API bool srslte_ue_dl_decode_phich(srslte_ue_dl_t *q,
